Fixed width, height, padding issues
Altered the width, height, padding, margin, etc for the img in order to
line it up with the rest of the content. Also did the same thing for the
nav section but still need to fix the issue with "Header Logo" that is
causing it to be offset. Seems to be a padding-left issue somewhere
but I can't find what's causing it to be so much further along the page
than the rest of the content. Unable to use flex-start because "justify-
content: space-between" is already being used.